Changes i128 to i64 in part 2

+++ src/problems/problem_13.rs
@@ -25,19 +25,19 @@ pub fn find_earliest_bus(input: &[u8]) -> u32 {
     earliest_bus.1 * (earliest_bus.0 - earliest_departure)
 }
 
-pub fn find_sequential_arrivals(input: &[u8]) -> i128 {
+pub fn find_sequential_arrivals(input: &[u8]) -> i64 {
     // For this problem the first line is irrelevant
     let (_, read_count) = lexical::parse_partial::<u32, _>(&input[..]).unwrap();
     let mut input = &input[read_count + 1..]; // Make sure we skip the line break
 
     let mut place_idx = 0;
-    let mut bus_ids: Vec<i128> = vec![];
-    let mut time_differentials: Vec<i128> = vec![];
+    let mut bus_ids: Vec<i64> = vec![];
+    let mut time_differentials: Vec<i64> = vec![];
     while let Some(first_char) = input.first() {
         if *first_char == b'x' {
             input = &input[2..]; // Make sure we also skip the comma
         } else {
-            let (bus_id, read_count) = lexical::parse_partial::<i128, _>(&input[..]).unwrap();
+            let (bus_id, read_count) = lexical::parse_partial::<i64, _>(&input[..]).unwrap();
             bus_ids.push(bus_id);
             time_differentials.push(place_idx);
             input = &input[read_count + 1..]; // Next character guaranteed to be comma to skip
@@ -45,7 +45,7 @@ pub fn find_sequential_arrivals(input: &[u8]) -> i128 {
         place_idx += 1;
     }
 
-    bus_ids.iter().product::<i128>() - chinese_remainder_theorem::<i128>(&time_differentials[..], &bus_ids[..]).unwrap()
+    bus_ids.iter().product::<i64>() - chinese_remainder_theorem::<i64>(&time_differentials[..], &bus_ids[..]).unwrap()
 }
 
 #[cfg(test)]
